requirements.txt
streamlit==1.32.0
requests==2.31.0
openai>=1.50.0
python-dotenv==1.0.0
pyatlan>=0.8.0 